+++ Yeti/Controllers/Sidebar/SidebarVC+Actions.h
@@ -20,6 +20,22 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)didTapSettings;
 
+#pragma mark - Common Action Handlers
+
+- (BOOL)feedCanShowExtraShareLevel:(Feed * _Nonnull)feed;
+
+- (void)shareFeedURL:(Feed * _Nonnull)feed indexPath:(NSIndexPath * _Nonnull)indexPath;
+
+- (void)shareWebsiteURL:(Feed * _Nonnull)feed indexPath:(NSIndexPath * _Nonnull)indexPath;
+
+- (void)feed_didTapShare:(Feed * _Nonnull)sender indexPath:(NSIndexPath * _Nonnull)indexPath;
+
+- (void)feed_didTapMove:(Feed * _Nonnull)sender indexPath:(NSIndexPath * _Nonnull)indexPath;
+
+- (void)confirmFeedDelete:(Feed * _Nonnull)feed completionHandler:(void(^ _Nullable)(BOOL actionPerformed))completionHandler;
+
+- (void)confirmFolderDelete:(Folder * _Nonnull)folder completionHandler:(void(^ _Nullable)(BOOL actionPerformed))completionHandler;
+
 @end
 
 NS_ASSUME_NONNULL_END